refactor(framework): Migrate event log from middleware to dependency
